fix(lint): move test-only glob cache helpers to _test.go
staticcheck reported globCacheLen and resetGlobCache as unused (U1000)
because they only exist for tests. Move them to a _test.go file so
they're excluded from production analysis.
